address-model compilation property value is now determined by the processor architecture and only for Android
If Boost is being compiled in a MacOS machine, using Darwin toolset, the address-model property is assigned with 32_64 to build a binary that can run both in 32-bit and 64-bit modes, when a 64-bit binary is needed (it won't accept just 64).
This is necessary only when the binaries required are 64-bit, instead of the 32-bit that is the default.

For darwin/32-bit or any other environment, the variable ADDRESS_MODEL is used to pass as a value (32 or 64), according to the processor architecture.
This requires confirmation by testing it other machines (Linux, Windows, etc.)