Treat root block in branch as any other block but with locator = Hash::null()
On one hand this simplified storing and fetching it from Branch, but it
also helps in that the merkle root can now be used as an identifier of
the content in the branch. Previously this wasn't the case because the
root block ID wasn't part of the merkle root calculation (it still could
have been done by creating another entry in the table equal to
HASH(merkle_root ++ root_block_id), but for now the implemented approach
wins in simplicity).