Stop using the `let _ = ...` construct. Use `unwrap_or(())` instead
It can cause subtle errors. Consider this example:

    fn foo() -> Result<()> { ... }

    // ...

    let _ = foo(); // we don't care about the error here

This works, but say we change `foo` to be `async` now:

    async fn foo() -> Result<()> { ... }

    // ...

    let _ = foo();

We think we are ignoring `Result` there, but we are actually ignoring `Future` which means the call to `foo` now silently does nothing.

If we instead use:

    foo().unwrap_or(())

Then the change to `async` would cause compile error which would remind us to add `await`.