Use BlobId to identify blobs in a directory
The reason for the change is that we'll soon want to keep multiple
versions of an entry in a directory. Because of this, when merging
branches, an entry could be forced to change it locators (because the
seq entry number could already be taken). Doing so would make mergning a
non commutative operation.