Have directory store multiple concurrent versions of each directory
Also removed a lot of code that was primarily working under the
assumption that there is only a single branch with a single version of
each directory entry. This code will need to be re-written together with
tests once the multi-version-entries refactor is finished and version
vectors are added to each directory entry.