Fix deadlock when accessing and modifying a directory concurrently
The reason for the deadlock was that the reader tried to open a
directory read lock in the "root to child" order but when the writer -
when flusing - was acquiring write locks in the "child to root" order.

The solution presented in this commit removes the necessity to lock the
child directory while the current directory is also locked during the
`open` operation.