Fix race condition in Network
Previously, registering a repository with the network worked like this:

1. Acquire lock to the message_brokers container
2. Create the repository link in all existing message brokers
3. Release the message_brokers lock
4. Acquire lock to registry
5. Insert the registration into the registry
6. Release the registry lock

What could happen is that a new connection and thus new message broker might have been created in between steps 3 and 4 which means the repository link would never get created because in step 2 the broker doesn't exists yet and at the time the broker is create the registration is not inserted into the registry.

Fix is to put both message_brokers and registry behind a single lock.