Use db::Connection (almost) everywhere
Turns out `db::Transaction` derefs to `db::Connection` so we can avoid generics (`Executor` and `Acquire` traits) while keeping the flexibility of being able to pass either connection or transaction to functions. One disadvantage is that we can no longer pass `Pool` directly and must call `acquire` (or `begin`) explicitly now. We can probably live with it. On the other hand, the explicitness is actually kind of nice.