Workaround to rupnp::discover function
This function tries to probe devices as it sees them from the
ssdp_client::search command. Problem is when one of the devices can't be
connected to (in my case the IP was on another subnet), then it'll take
~90 seconds to timeout before other devices can be probed.

In this commit we're probing each device in a separately spawned
coroutine and thus avoiding the problem.