Avoid processing nodes that don't contain any new information
Previously we processed a node as long as its summary was more up to date compared to the corresponding node already stored in our index. After this commit, we only process it if it's also more up to date compared to the last time we received it. This dramatically reduces the ammount of messages sent in some cases.