Fix incorrect select! macro usage
Example:

```rust
loop {
    select! {
        Some(item) = rx.recv() => { ... }
        _ = time::timeout(...) => { ... }
        else => break,
    }
}
```

This is wrong. The intent was that the `else` branch is hit when `rx.recv` returns `None`, but that's not what happens. Instead the next branch is tried and because it matches everything (`_`) it is always executed when ready. The correct thing to do instead is this:

```rust
loop {
    select! {
        item = rx.recv() => {
            if let Some(item) = item {
                ...
            } else {
                break;
            }
        }
        _ = time::sleep(...) => { ... }
    }
}
```