Remove Pool::detach
We were only using it in tests to detach a connection from the pool and then drop the pool, keeping only the connection. Turns out it's possible to simply `acquire` a connection and then drop the pool. The connection will remain open. Thus separate `detach` is unnecessary for our purposes.