Avoid some unnecessary ChildNodes requests
When receving InnerNodes response we compare the received nodes against the nodes we already have and accept only the ones that are more up-to-date. There was a bug in this comparison code where we would try to load the local nodes with the same parent_hash as the received nodes. But if the remote replica performed any change, the parent_hash would always be different and so there would be no local nodes having the same parent_hash and the comparison would conclude that all the received nodes are more up-to-date than any we have, resulting in unnecessary further requests being sent.

The fix is to load the local nodes with the same hashes (not parent_hashes) as the received nodes.