Replace Summary::{missing_blocks -> block_presence}
Instead of storing the number of missing blocks per node, we now only store a simple 3-state value indicating whether we have all, some or none of the blocks under than node. Even though this is seemingly less information that previously, it actually allow us to make better decision whether to descend into the child node or not than before. This is because
    a) we didn't really need the actual number of missing blocks to make the decision, only whether there is all, some or none of them.
    b) previously it was hard to tell the case where all blocks are missing from the one where only some are. Now it's easy.