Fix the FFI repository_set_read_and_write_access function
Also adds an optional `local_old_rw_password` which, if present, will
serve to retrieve the old writer ID and store it back to the metadata
table encrypted with the new password. If the old password is not
present, a new writer ID shall be generated. The second use case may be
useful for when the user did not have a prior write access but now
obtained a new RW share token.