Fix message reorder bug in the dispatcher
Imagine two `ContentStreams` `s1` and `s2` working on different channels
`ch1` and `ch2` respectively.

`s1` receives a message `m1` for `ch2` and so pushes it back to the
queue.  `s2` is in the `select!` and receives a message `m2`. The next
time `ContentStream::recv()` is called, `s2` will check the queues and
pops `m1` from it.