[Android] preview file improved
When previewing a file, we now check for the mime type to determine if it is a valid file type before attempting to create the intent.

If the extensions is valid (mime type is not null), then we start the activity using the intent, and check for an ActivityNotFoundException if there is not app for it.

(This was the most effective way to do it, since using intent.resolveActivity sometimes returns false negatives)

In any of those cases, or a successful intent initiation, we return a string to show a message to the user in case if was not possible to do it.