Remove ReceiveFilter (not needed)
Whenever a `RootNode` or `InnerNodes` response is received, it's compared against the corresponding root/inner nodes that already exists locally in order to determine whether a followup requests should be sent. This means that if a response that indicates the peer has some blocks we don't is received multiple times, we would have to follow up every time until we actually receive those blocks (because only then are the node summaries updated). This would be very innefficient and so `ReceiveFilter` was introduced to filter those duplicate responses out. However, after the introduction of the `BlockOffer` response, the peers no longer send a `RootNode` response every time they receive new blocks. They send a `BlockOffer` response instead and the followup to that (the `Block` request) is sent only if one is not already in flight. Thus `ReceiveFilter` is no longer needed. This change has no performance impact as confirmed by benchmarks.