Clarify how the seed is used to create the write key
I tripped over the word _interpret_ here on step 2, since the private key is a scalar i _guess_ you _could_ use it directly but that would skip the SHA512 hashing and clamping that is given in [key generation section of RFC803](https://www.rfc-editor.org/rfc/rfc8032#section-5.1.5). 

I think using the word _derive_ instead makes it clearer that this is more involved and less ambigous with the use of the word on step 4.